[ttpe] add optional argument specifying test to rebaseline and mkreview scripts

+++ ttt-ttpe/src/tools/scripts/rebaseline.sh
@@ -2,10 +2,21 @@
 cd ../..
 SRCDIR=src/test/resources/com/skynav/ttpe/app
 RUNDIR=run/test
-for f in $SRCDIR/test-*.xml
+if [ $# -gt 0 ]; then
+  p=$1
+else
+  p=test-
+fi
+for f in $SRCDIR/$p*.xml
 do
-  echo Rebaselining $f ...
-  cp $f $RUNDIR/test.xml
-  mvn exec:java
-  cp $RUNDIR/out/test.zip ${f%.xml}.expected.zip
+  if [ ! -e $f ]; then
+    continue
+  elif [ "$f" == 'test.config.xml' ]; then
+    continue
+  else
+    echo Rebaselining $f ...
+    cp $f $RUNDIR/test.xml
+    mvn exec:java
+    cp $RUNDIR/out/test.zip ${f%.xml}.expected.zip
+  fi
 done
